manta-connector: add private workqueue
There is a function change_charger_status(), which can
be called from the both contexts: worker and interrupt.
This function uses mutex to synchronize the access.
Moreover, inside critical setion, it schedules a task
to generic nrt workqueue and waits for its completion.

The problem occured, when change_charger_status() is called
for worker context, and is preempted by interrupt. So the
following condition occured:
1. Mutex is aquired by worker;
2. Inside critical section, a job is scheduled to low-priority
   workqueue;
3. Interrupt thread is blocked on mutex mentioned above.

Due to interrupt thread has higher priority, the nrt workqueue
is never scheduled, so a deadlock occurred.

This patch adds a private high-priority workqueue which can't
be preempted by interrupts.

Signed-off-by: Alexander Smirnov <asmirnov@ilbers.de>